requirements.txt
# Quantitative Research Bot - Required Libraries
# Install with: pip install -r requirements.txt

# A. Data Handling (Core)
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<2.2.0
pyarrow>=14.0.1,<16.0.0
fastparquet>=2024.5.0
python-snappy>=0.7.1
zstandard>=0.22.0

# B. API Clients
ibapi>=10.41.0

# C. Configuration & Utilities
PyYAML>=6.0
rich>=13.0.0

# D. Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# E. Machine Learning
torch>=2.0.0
scikit-learn>=1.3.0

# F. Utilities
tqdm>=4.65.0
ib_insync>=0.9.86
